% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{print.palette}
\alias{print.palette}
\title{Prints an example of the chosen palette}
\usage{
\method{print}{palette}(x, ...)
}
\arguments{
\item{x}{Name of album - NEED UPDATE}

\item{...}{Additional arguments, currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
NEEDS UPDATE
Use \code{\link[=psb_palette]{psb_palette()}} to get palettes of the desired length.
}
