% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psb.R
\name{psb_palette}
\alias{psb_palette}
\title{Generate PSB themed palette}
\usage{
psb_palette(x)
}
\arguments{
\item{x}{name of psb album NEED UPDATE}
}
\description{
Use \code{\link[=psb_palette]{psb_palette()}} to get palettes of the desired length.
}
