% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{psb_palettes}
\alias{psb_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 15.
}
\usage{
psb_palettes
}
\description{
Use \code{\link{psb_palette}} go get palettes of desired length
}
\keyword{datasets}
