% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{print.palette}
\alias{print.palette}
\title{Prints an example of the chosen palette}
\usage{
\method{print}{palette}(x, ...)
}
\description{
NEEDS UPDATE
Use \code{\link[=psb_palette]{psb_palette()}} to get palettes of the desired length.
}
